+++ src/main/kotlin/app/revanced/patches/reddit/customclients/baconreader/api/SpoofClientPatch.kt
@@ -14,9 +14,7 @@ import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
 
 @Patch(
     name = "Spoof client",
-    description = "Spoofs the client in order to allow logging in. " +
-            "The OAuth application type has to be \"Installed app\" " +
-            "and the redirect URI has to be set to \"http://baconreader.com/auth\".",
+    description = "Restores functionality of the app by using custom client ID's.",
     compatiblePackages = [
         CompatiblePackage("com.onelouder.baconreader"),
         CompatiblePackage("com.onelouder.baconreader.premium")